Refactor formatting, fix idle timer, add ZenMode focus trap
