+++ frontend/src/utils/formatting.ts
@@ -1,0 +1,26 @@
+export function formatDateShort(isoString: string): string {
+  try {
+    return new Date(isoString).toLocaleDateString('en-US', {
+      month: 'short', day: 'numeric', year: 'numeric',
+    })
+  } catch { return '' }
+}
+
+export function formatDateLong(isoString: string): string {
+  try {
+    return new Date(isoString).toLocaleDateString('en-US', {
+      month: 'long', day: 'numeric', year: 'numeric',
+    })
+  } catch { return '' }
+}
+
+export function formatRelativeTime(timestamp: number): string {
+  const diff = Math.floor(Date.now() / 1000) - timestamp
+  if (diff < 3600) return `${Math.floor(diff / 60)}m ago`
+  if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`
+  return `${Math.floor(diff / 86400)}d ago`
+}
+
+export function getDomain(url: string): string {
+  try { return new URL(url).hostname.replace(/^www\./, '') } catch { return '' }
+}
